requirements.txt
django >= 2.2,<2.3
django-debug-toolbar
django-jinja
djangorestframework
psycopg2
psycopg2-binary
requests
pyyaml

# channels
django-channels
channels-redis